+++ packages/form-field/src/FormField/FormField.styles.ts
@@ -1,9 +1,15 @@
 import { css } from '@leafygreen-ui/emotion';
-import { Theme } from '@leafygreen-ui/lib';
-import { palette } from '@leafygreen-ui/palette';
-import { BaseFontSize, Size, spacing, typeScales } from '@leafygreen-ui/tokens';
+import {
+  BaseFontSize,
+  Size,
+  spacing,
+  typeScales,
+  Variant,
+} from '@leafygreen-ui/tokens';
 
-export const getFontSize = ({
+import { FormFieldState } from './FormField.types';
+
+export const getFontSizeStyles = ({
   baseFontSize,
   size,
 }: {
@@ -26,10 +32,22 @@ export const getFontSize = ({
 
   if (size === Size.Large) {
     return css`
-      font-size: 18px;
-      line-height: 24px;
+      font-size: ${typeScales.large.fontSize}px;
+      line-height: ${typeScales.large.lineHeight}px;
     `;
   }
+};
+
+export const convertFormFieldStateToIconVariant = (state: FormFieldState) => {
+  if (state === FormFieldState.Error) {
+    return Variant.Error;
+  }
+
+  if (state === FormFieldState.Valid) {
+    return Variant.Success;
+  }
+
+  return Variant.Primary;
 };
 
 export const textContainerStyle = css`
@@ -40,28 +58,3 @@ export const textContainerStyle = css`
 export const marginBottom = css`
   margin-bottom: ${spacing[100]}px;
 `;
-
-export const stateFeedbackContainerStyle = css`
-  margin-top: ${spacing[100]}px;
-  display: flex;
-  align-items: center;
-  gap: ${spacing[100]}px;
-`;
-
-export const errorIconStyles: Record<Theme, string> = {
-  [Theme.Light]: css`
-    color: ${palette.red.base};
-  `,
-  [Theme.Dark]: css`
-    color: ${palette.red.light1};
-  `,
-};
-
-export const validIconStyles: Record<Theme, string> = {
-  [Theme.Light]: css`
-    color: ${palette.green.dark1};
-  `,
-  [Theme.Dark]: css`
-    color: ${palette.green.base};
-  `,
-};