+++ packages/form-field/src/FormField/useFormFieldProps.ts
@@ -1,5 +1,7 @@
 import { useIdAllocator } from '@leafygreen-ui/hooks';
 
+import { LGIDS_FORM_FIELD } from '../constants';
+
 import { FormFieldProps, FormFieldState } from './FormField.types';
 
 export interface FormFieldInputElementProps {
@@ -15,7 +17,7 @@ export interface FormFieldInputElementProps {
 export interface FormFieldElementProps {
   labelId: string;
   descriptionId: string;
-  errorId: string;
+  feedbackId: string;
   inputId: string;
   inputProps: FormFieldInputElementProps;
 }
@@ -28,20 +30,21 @@ export const useFormFieldProps = ({
   disabled,
   ...rest
 }: Partial<FormFieldProps>): FormFieldElementProps => {
-  const labelId = useIdAllocator({ prefix: 'lg-form-field-label' });
+  const labelId = useIdAllocator({ prefix: LGIDS_FORM_FIELD.label });
   const descriptionId = useIdAllocator({
-    prefix: 'lg-form-field-description',
+    prefix: LGIDS_FORM_FIELD.description,
   });
-  const errorId = useIdAllocator({ prefix: 'lg-form-field-description' });
-  const generatedInputId = useIdAllocator({ prefix: 'lg-form-field-input' });
+  const feedbackId = useIdAllocator({ prefix: LGIDS_FORM_FIELD.feedback });
+  const generatedInputId = useIdAllocator({ prefix: LGIDS_FORM_FIELD.input });
   const inputId = id ?? generatedInputId;
 
   const hasError = state === FormFieldState.Error;
+  const hasFeedback = state !== FormFieldState.None;
 
   const ariaLabelledby = label ? labelId : rest['aria-labelledby'];
   const ariaLabel = label || ariaLabelledby ? undefined : rest['aria-label'];
   const ariaDescribedby = `${description ? descriptionId : ''} ${
-    hasError ? errorId : ''
+    hasFeedback ? feedbackId : ''
   }`.trim();
   const ariaInvalid = rest['aria-invalid'] ?? hasError;
 
@@ -58,7 +61,7 @@ export const useFormFieldProps = ({
   return {
     labelId,
     descriptionId,
-    errorId,
+    feedbackId,
     inputId,
     inputProps,
   };