+++ packages/tokens/src/color/lightModeColors.ts
@@ -1,6 +1,11 @@
 import { palette } from '@leafygreen-ui/palette';
 
-import { State, Variant } from './color.types';
+import {
+  ModeColorRecord,
+  State,
+  Variant,
+  VariantColorRecord,
+} from './color.types';
 
 const { black, blue, gray, green, red, white, yellow } = palette;
 
@@ -45,9 +50,9 @@ const lightModeBackgroundColors = {
     [State.Hover]: gray.light2,
     [State.Focus]: gray.light2,
   },
-};
+} as const satisfies VariantColorRecord;
 
-const lightModBorderColors = {
+const lightModeBorderColors = {
   [Variant.Primary]: {
     [State.Default]: gray.base,
     [State.Hover]: gray.base,
@@ -73,7 +78,7 @@ const lightModBorderColors = {
     [State.Hover]: gray.light1,
     [State.Focus]: gray.light1,
   },
-};
+} as const satisfies VariantColorRecord;
 
 const lightModeIconColors = {
   [Variant.Primary]: {
@@ -116,9 +121,9 @@ const lightModeIconColors = {
     [State.Hover]: gray.base,
     [State.Focus]: gray.base,
   },
-};
+} as const satisfies VariantColorRecord;
 
-export const lightModeTextColors = {
+const lightModeTextColors = {
   [Variant.Primary]: {
     [State.Default]: black,
     [State.Hover]: black,
@@ -149,11 +154,21 @@ export const lightModeTextColors = {
     [State.Hover]: gray.base,
     [State.Focus]: gray.base,
   },
-};
+  [Variant.Placeholder]: {
+    [State.Default]: gray.base,
+    [State.Hover]: gray.base,
+    [State.Focus]: gray.base,
+  },
+  [Variant.Link]: {
+    [State.Default]: blue.base,
+    [State.Hover]: blue.base,
+    [State.Focus]: blue.base,
+  },
+} as const satisfies VariantColorRecord;
 
 export const lightModeColors = {
   background: lightModeBackgroundColors,
-  border: lightModBorderColors,
+  border: lightModeBorderColors,
   icon: lightModeIconColors,
   text: lightModeTextColors,
-};
+} as const satisfies ModeColorRecord;