+++ packages/typography/src/Label/Label.tsx
@@ -12,9 +12,8 @@ import { LGIDS_TYPOGRAPHY } from '../constants';
 import { useUpdatedBaseFontSize } from '../utils/useUpdatedBaseFontSize';
 
 import {
-  disabledLabelColorStyle,
-  labelColorStyle,
-  labelStyle,
+  getDisabledLabelColorStyle,
+  getLabelStyles,
   labelTypeScaleStyles,
 } from './Label.styles';
 import { BaseLabelProps } from './Label.types';
@@ -38,10 +37,9 @@ export const Label = Polymorphic<BaseLabelProps>(
       <Component
         data-lgid={dataLgId}
         className={cx(
-          labelStyle,
-          labelColorStyle[theme],
+          getLabelStyles(theme),
           labelTypeScaleStyles[baseFontSize],
-          { [disabledLabelColorStyle[theme]]: disabled },
+          { [getDisabledLabelColorStyle(theme)]: disabled },
           className,
         )}
         {...rest}