+++ chat/title-bar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lg-chat/title-bar
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [dfd6972c]
+  - @leafygreen-ui/typography@19.0.0
+
 ## 2.0.1
 
 ### Patch Changes