+++ packages/typography/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @leafygreen-ui/typography
 
+## 19.1.0
+
+### Minor Changes
+
+- c406ab85: [LG-2930](https://jira.mongodb.org/browse/LG-2930)
+
+  - Updates typography components to use tokens where possible
+
+### Patch Changes
+
+- Updated dependencies [c406ab85]
+  - @leafygreen-ui/tokens@2.6.0
+
+## 19.0.0
+
+### Major Changes
+
+- dfd6972c: - Fixes the `Link` component so that:
+  - links are inline with the text preceding it
+  - The Arrow Icon is inline with the last line of text. Previously it rendered in a column next to the text because of `display: inline-flex`
+  - Adds generated stories for each `Typography` component
+
 ## 18.4.0
 
 ### Minor Changes